Powerups!
The major feature included in this update are the powerups. They are in
the game but don' yet affect the gameplay (they get drawn, but don't
provide any other functions yet)

Also added the resources (theme.res) and changed the build settings.
(Now has a proper icon displayed on the phone, instead of the CNO
default)

Other updates can be read in the readme file! :)